fix: update snapshot can't propagate when using efcore
When add a new snapshot to an exist document, because the snapshot' primary is a user set entity id(not zero), efcore will treat this newly snapshot as a disconnected existed document, thus could not persist successfully. To solve this, add a trivial GUID property as primary key.

The related unit tests were refactored to meet AAA.